requirements.txt
# SENTINEL Community Edition - Dependencies
# Pinned versions for reproducibility (2026-01-02)

# Core
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# NLP / ML
sentence-transformers>=3.3.0
transformers>=4.47.0
torch>=2.5.0

# Detection
presidio-analyzer>=2.2.355
presidio-anonymizer>=2.2.355
yara-python>=4.5.2

# Topological Data Analysis
gudhi>=3.10.1
numpy>=2.1.0
scipy>=1.14.0

# VLM Protection
Pillow>=11.0.0
pytesseract>=0.3.13

# CLIP (cross-modal)
# openai-clip>=1.0.1  # Optional, install separately if needed

# API
httpx>=0.28.0
python-multipart>=0.0.18
openai>=1.0.0

# Utilities
pyyaml>=6.0.2
python-dotenv>=1.0.1
structlog>=24.4.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0

# Benchmarks
matplotlib>=3.9.0
plotly>=5.24.0
datasets>=3.1.0